INTERNAL: ath10k: Fix rate table updation in tx stats
The index for updating rate table tx stats is
calculated using the bandwidth value. The bandwidth
values do not map correctly with the bandwidth values
shown in the rate table.

Correct the bandwidth value calculation which is used
to calculate the index for rate table updation for tx stats.

CRs-Fixed: 2407286
Change-Id: Id25a4c48379f680d247795473e6b8b1dd5d69565
Signed-off-by: Rakesh Pillai <pillair@codeaurora.org>