INTERNAL: ath10k: Fix NSS tx stats for legacy rates
The ratecode field given by firmware in the tx stats
gives the details of NSS and MCS, if the preamble
specifies that it is HT/VHT. In case of OFDM/CCK, as
specified by the preamble, the ratecode indicates the
legacy rate instead of NSS/MCS data. For OFDM/CCK cases,
the NSS should always be 1, i.e. 1x1 mode for 11g and 11b.

Currently the legacy rate is misinterpreted as NSS & MCS, leading
to incorrect updation of tx stats corresponding to NSS.

Fix the nss value to 1 for 11b and 11g modes by checking
the preamble indication in the ratecode provided by the firmware.

CRs-Fixed: 2407311
Change-Id: Id37e3565de0f7c5dcfac5d97948a1c760fba1a34
Signed-off-by: Rakesh Pillai <pillair@codeaurora.org>