driver: mmc: mmci_qcom_tuning: Add 1ms delay for proper tuning
This change will add 1ms delay for proper tuning. for proper
tuning wait for 146 MCLK cycles for the card to send out the data
and thus move to TRANS state. As the MCLK would be minimum 200MHz when
tuning is performed, we need maximum 0.73us delay. To be on safer side
1ms delay is given.

Without this change we can see tuning failure occasionally.

error:
[3.911363] mmc0: tuning execution failed
[3.913721] mmc0: error -11 whilst initialising SD card

Change-Id: I50b0e6ed00634627e18db8939267dbc984896f2d
Signed-off-by: Md Sadre Alam <mdalam@codeaurora.org>