remoteproc: Disable autoboot of remote processor.
The current framework will automatically try to boot the remote
processor if rproc->auto_boot variable is set. However the start and
stop sequence of the Q6 need to be controlled by the host driver.
For supporting this, we are disabling autoboot of remote processor
by setting rproc->auto_boot to false. This way host driver can make use
of rproc_boot and rproc_shutdown api's to change the states of remote
processor(start or stop).

Change-Id: I6d3340617dbe9297fa15672ec3291c1737079c27
Signed-off-by: Nikhil Prakash V <nprakash@codeaurora.org>