iommu: enable add_device callback on BUS_NOTIFY_ADD_DEVICE event
Devices behind the IOMMU should be attached to IOMMU domains before the
device probe. IOMMU driver registers the notifiers on BUS_NOTIFY_ADD_DEVICE
event for IOMMU add_device.  During the of_device_add, before the device
probe, this notifier chain is triggered so that the device is properly
attached to IOMMU domain.

Fixes: I9d056ab0512 (drivers: iommu: Changes to enable iommu for ipq8074)
Change-Id: If26e20d772c431e054bcbc5eb356e08fd665af03
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>