usb: dwc3: Continue using qcom,dwc3 compatible for of-simple driver
In the dwc3-qcom upstream driver, they have removed qcom,dwc3
compatible from of-simple driver and started to use the same
for dwc3-qcom driver. Since, in most of the IPQ platforms,
we continue to use the of-simple driver, we have updated the
compatible in of-simple driver and use new compatible for
dwc3-qcom driver.

Change-Id: I03eae255f099a9bc80ad359f5760e316a0f11713
Signed-off-by: Balaji Prakash J <bjagadee@codeaurora.org>