arm64: dts: Update memory layout for ipq6018
This patch updates the memory layout for ipq6018
platform.

Taking in consideration possibilities of growth
of the memory required for LPASS/WCSS in future,
this change was done to update the memory layout.

Also with this layout, tzapps, uboot, sbl can be
removed if in case its not needed without having
to change the start address of other partitions.

Change-Id: I447b377eaecfe659c0313a51aa46aa837c2e8bff
Signed-off-by: speriaka <speriaka@codeaurora.org>