drivers: firmware: scm: get the SCM CMD ID for secure sysupgrade from DTS
For IPQ6018 chipset, the SCM CMD ID for secure sysupgrade is 0x1F, which
is different from the existing one (0x15) used in the IPQ807x and other
legacy chipset.

With this patch, if the SCM CMD ID is defined in DTS the same will be
used, else the existing one (0x15) will be used.

Change-Id: I1150aeb3f8e6c93b204d2073a43c68fa9da0d7bc
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>