rproc: q6v5: Add emulation start to rproc ops before rproc alloc
When emulation flag is set we are updating the rproc ops structure
to use the emulation start API. Currently this is done after
rproc_alloc succeeded.

When the rproc_ops structure passed to rproc_alloc API, it duplicates
the ops structure and keeps a copy of it. Since we are assigning
the start API after rproc_alloc the emu start will not get updated.

This patch handles the situation by changing the rproc ops structure
before it is being passed to the rproc_alloc.

Change-Id: I4131e1811f5030897e61f1ba03d9319d8629f966
Signed-off-by: Sivaprakash Murugesan <sivaprak@codeaurora.org>