ipq807x: arm: Text offset fix for 256M profile
The default text offset, 0x208000, makes kernel consider the DDR
start address as 0x4100_0000.
              DDR start address = LOAD Address - Text offset
                    Load address = 0x41208000
                    Default Text offset = 0x208000
              DDR start address = 0x41208000 - 0x208000
                               --> 0x41000000
This is fine in normal profile as the first 16MB are reserved memory.
But when it comes to 256MB profile, only first 8MB are reserved, which
causes the 8MB of DDR to be unused.
To fix this a new config is added, enabling the config will make the
text offset as 0x1208000, which will enable the kernel to recognize
the DDR start address as 0x4000_0000.

Page Table base addr calculation changes:-
The page table base address (swapper_pg_dir) is calculated as
   swapper_pg_dir = Physical address + TEXT_OFFSET - pg dir size
   The original text offset was enough to be expressed as an
   immediate constant in an add instruction. The new text offset
   0x1208000 can't be expressed as an immediate constant. Hence
   its loaded to a register and added in pgtbl macro used in
   __create_page_tables.

Memblock changes:-
When there is a reserved region at the start of DDR, then it is
removed from the memblock DDR type. Later, the first used DDR
addresses is taken as the start of DDR. This is wrong, this
causes the bootmem_init to use a wrong offset. This patch
adds a new variable for recording the DDR start address.

Change-Id: I5feeab0538fa72c4a83c6d072c174c0adbcfcfb5
Signed-off-by: Manoharan Vijaya Raghavan <mraghava@codeaurora.org>