bridge: Add support for IEEE 802.11 Proxy ARP for IPv6
This is an IPv6 extension of commit 958501163ddd ("bridge: Add support
for IEEE 802.11 Proxy ARP") for 4.4 kernel. The IEEE 802.11 Proxy ARP
feature is defined in IEEE Std 802.11-2012, 10.23.13. It allows the AP
devices to keep track of the hardware-address-to-IP-address mapping of
the mobile devices within the WLAN network.

The AP will learn this mapping via observing NS/NA frames (this part is
implemented in user space, e.g., in hostapd). When a request for such
information is made (i.e., Neighbor Solicitation), the AP will respond
on behalf of the associated mobile device. In the process of doing so,
the AP will drop the multicast request frame that was intended to go out
to the wireless medium.

To meet the Hotspot 2.0 requirement of using the target devices MAC
address as the sender's MAC address in the NA frame (when AP replies on
behalf of an associated station), the full link layer header needs to be
built with a custom function instead of going through the existing
ndisc.c implementation.

Change-Id: I07f16a1dc11b873260782c77e51044755800523c
Signed-off-by: Jouni Malinen <jouni@codeaurora.org>
Signed-off-by: Sathishkumar Muruganandam <murugana@codeaurora.org>