remoteproc: q6v6: update smp2p stop bit when force stop is false
smp2p stop message is used to gracefully shutdown the remote
processor. Currently this will be updated only when the
remote processor is crashed, so the graceful shutdown will not
take place in normal shutdowns.

This change will send the smp2p stop message on normal shutdowns,
so that the remote processor can cleanup the resources.

Change-Id: Iacb0f118bd6430fc5c8d6e067f0d9d157fdb6236
Signed-off-by: Sivaprakash Murugesan <sivaprak@codeaurora.org>