[arm64][dts] Fix build warning
We need to explicit mention address cell and size cell for nodes with
"reg" property. This is required because by default for 64 bit build,
"reg" property address cell size width will be taken as 2 words
which will mismatch with the usage of the following type:

reg = <0x39800000 0x7ffff>; Here the address-cell size is 1 but by default
it will take it as 2 for 64 bit build.

Change-Id: I633450662438879e2ce573505e4102df629f087f
Signed-off-by: Sourav Poddar <souravp@codeaurora.org>