minidump: Support invalidating Crashdump TLVs
Since the size of the Global crashdump buffer is
limited to 16K, a maximum of approx. 862 TLVs can
be supported. If TLV # exceeds 862, the minidump
code exits gracefully, but the dumps collected may
be incomplete.
This change brings in support to invalidate “outdated”
TLVs in the global crashdump buffer and replace TLVs
rather than always append TLVs to the buffer.
This will help support 862 unique TLVs.

Change-Id: I47906d97aebaf4b2a7bc2f4bf3b3ce126b9ade27
Signed-off-by: Gitanjali Krishna <gitanjal@codeaurora.org>