sdhci: msm: Use separate notifier block for reboot and panic
Same notifier block structure was used for reboot and panic
notifier list. The notifier block structure is first registered
with panic list and then to reboot list. The next pointer of
this notifier block gets updated with the upcoming reboot
notifier being registered. So, reboot notifiers are getting
called during the execution of panic_handler_list.

Change-Id: I432ddebf1606014a3bdbcf761657548f23c64bcb
Signed-off-by: Balaji Prakash J <bjagadee@codeaurora.org>