drivers: iommu: add debug mechanism to track iommu domains
Currently, our IOMMU driver lacks the certain debug information. One of
them is tracking the list of iommu domains. This change tries to address
that by creating the list and on every __iommu_attach_device and
iommu_domain_free, we are updating the domain and its group information to
the list.

This will be helpful to dump the iommu domain and its information from
the ramdump as well.

Example Output:

IOMMU Context: None attached. Domain: ab00000.adsp_rproc[L2 cache redirect for page tables is OFF]
[VA Start -- VA End] [Size] [PA Start -- PA End] [Attributes][Page Table Entry Size] [Memory Type] [Shareability] [Non-Executable]

0x0--0x3ffff [0x3ffff] [UNMAPPED]
0x40000--0x40fff [0x1000] A:0x40000--0x40fff [0x1000] [R/W][4K] [Non-Cached] [Inner-Shareable] [False]
0x41000--0x41fff [0x1000] A:0x41000--0x41fff [0x1000] [R/W][4K] [Non-Cached] [Inner-Shareable] [False]
0x42000--0x42fff [0x1000] A:0x42000--0x42fff [0x1000] [R/W][4K] [Non-Cached] [Inner-Shareable] [False]
0x43000--0x43fff [0x1000] A:0x43000--0x43fff [0x1000] [R/W][4K] [Non-Cached] [Inner-Shareable] [False]

Change-Id: I4e18f96d27908c9c458c78783c76378b6bbc325f
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>