drivers: watchdog: qcom: add support to read the TCSR register from DTS
To handle the external watchdog, we are configuring the TCSR_WONCE
register with the appropriate value. TSCR_WONCE register to be used for
ipq60xx is different from the one used in ipq807x. So read the TSCR_WONCE
register from DTS, if not defined in DTS, ipq807x TCSR_WONCE register
will be used as default.

Also, this change takes care of the param ids for SCM IO operation which
should be SCM_VAL not SCM_RW.

Change-Id: I1bd08b14cc2890b91df08ab9942a95d7d3dd602a
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>