INTERNAL: soc: qcom: Add snapshot of the memory dump v2 driver
Currently, each dump table entry is being registered in several
places in the kernel. Add centralized support in memory dump
driver to parse the device tree to register and allocate dump
table entries. We rely on CMA memory pool to allocate memory for
the dump entries. Clients can either use the generic approach
or configure the dump table entries independently using the
existing framework.

This is a snapshot of the memory dump v2 driver from msm-4.14.

Change-Id: Ie5ff046348c0329b309cc865e3b0b50ff5dd85b7
Signed-off-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>