INTERNAL: Add support for miscellaneous debug features
INTERNAL: sched/rt: print RT tasks when RT throttling is activated

Existing debug prints do not provide any clues about which tasks
may have triggered RT throttling. Print the names and PIDs of
all tasks on the throttled rt_rq to help narrow down the source
of the problem.
This is a snapshot taken from msm-4.14.

INTERNAL: sched/rt: Add Kconfig option to enable panicking for RT
throttling

This may be useful for detecting and debugging RT throttling issues.
This is a snapshot taken from msm-4.14.

INTERNAL: sched: Introduce PANIC_ON_SCHED_BUG

Introduce PANIC_ON_SCHED_BUG to catch issues related to scheduler
like calling sleepable functions inside atomic context etc.
This is a snapshot taken from msm-4.14.

INTERNAL: mm: slub: panic for object and slab errors

If the SLUB_DEBUG_PANIC_ON Kconfig option is selected,
also panic for object and slab errors to allow capturing
relevant debug data.
This is a snapshot taken from msm-4.14.

INTERNAL: mm: Kconfig: Add DEBUG_SLUB_PANIC_ON option

Add the DEBUG_SLUB_PANIC_ON option to KCONFIG preventing
the existing defconfig option from being overwritten
by make config.
This will induce a panic if slab debug catches corruptions
within the padding of a given object.
The intention here is to induce collection of data
immediately after the corruption is detected with
the goal to catch the possible source of the corruption.
This is a snapshot taken from msm-4.14.

INTERNAL: lib: spinlock: Cause a watchdog bite on spin_dump

Currently we cause a BUG_ON once a spin_bug is detected, but
that causes a whole lot of processing and the other CPUs would
have proceeded to perform other actions and the state of the system
is moved by the time we can analyze it.
Provide an option to trigger  a watchdog bite instead so that we
can get the traces as close to the issue as possible.
This is a snapshot taken from msm-4.14.

Change-Id: I835f624b4e38a3efb7c2d0f7936472fed000295e
Signed-off-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>