FROMLIST: dt-bindings: soc: qcom: Add RSC power domain specifier
In addition to transmitting resource state requests to the remote
processor, the RSC is responsible for powering off/lowering the
requirements from CPUs subsystem for the associated hardware like
buses, clocks, and regulators when all CPUs and cluster is powered down.

The power domain is configured to a low power state and when all the
CPUs are powered down, the RSC can lower resource state requirements
and power down the rails that power the CPUs.

Add PM domain specifier property for RSC controller.

Change-Id: I0ba109ad473b40fcd7d19b34fba6627a963a9e21
Patch-mainline: https://lore.kernel.org/patchwork/patch/1114298
Cc: devicetree@vger.kernel.org
Signed-off-by: Maulik Shah <mkshah@codeaurora.org>