msm: disp: dpu: add support to access hw irqs regs depending on revision
Current code assumes that all the irqs registers offsets can be
access in all the hw revisions; this is not the case for some
targets that should not access some of the irq registers.
This change adds the support to selectively remove the irqs that
are not supported in some of the hw revisions.

Change-Id: I7fb1428b1573146065a198667e812e2002e59520
Signed-off-by: Shubhashree Dhar <dhar@codeaurora.org>