drivers: clk: qcom: ipq6018: add CLK_IGNORE_UNUSED flag
Currently, most of the clocks are turned on by the boot loaders. When
the clk_ignore_unused kernel command line is not passed, kernel
disables the unused clocks. Few clocks shouldn't be turned off, so add
the CLK_IGNORE_UNUSED flags to the required clocks to avoid the unexpected
behaviour.

Change-Id: I76688d408881a5759ef71a1692352b48f8c6ec60
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>