net: qualcomm: rmnet: Only set partial csum on coalesced descriptors
Only descriptors that were created as a result of coalescing (i.e. RSB,
RSC, or the rmnet_perf module) need to be set as partial checksum as the
L4 checksum in the headers will no longer be correct. Packets that were
not ever coalesced but contain valid checksums can simply be marked as
CHECKSUM_UNNECESSARY as they were previously.

    16.609699:   <6> Unable to handle kernel paging request.
    16.610794:   <2> Call trace:
    16.610799:   <2>  rmnet_frag_deliver+0x478/0x718
    16.610813:   <2>  rmnet_perf_core_flush_curr_pkt+0x12c/0x148
    16.610821:   <2>  rmnet_perf_core_desc_entry+0x12c/0x168
    16.610826:   <2>  rmnet_frag_ingress_handler+0x288/0x3c8
    16.610830:   <2>  rmnet_rx_handler+0x188/0x238
    16.610838:   <2>  __netif_receive_skb_core+0x444/0xb68
    16.610842:   <2>  process_backlog+0x170/0x390
    16.610846:   <2>  net_rx_action+0x134/0x548
    16.610853:   <2>  __do_softirq+0x1dc/0x384
    16.610858:   <2>  irq_exit+0xb4/0xb8

CRs-Fixed: 2529688
Change-Id: I5899eabdd033cb5bfedd65ad7fc970ae5ee8f920
Signed-off-by: Sean Tranchetti <stranche@codeaurora.org>