mhi: core: Fix for invalid ipcrouter packet.
The issue was with the xfer_len being passed to the client
notifier callback. In case of IPCR channel, they have two
properties auto-start and auto-queue. Due to this property,
the host will pre allocate buffer for xfer ring which the client
will later queue. However if there is more than one TRE, the xfer_len
was taken as the len of preallocated buffer, which is MHI_MAX_MTU
or 65535. However QRTR needs length to be byte aligned. So for
packets of len 65535 was dropped as invalid packet by QRTR
layer.

So added a fix to always get the xfer_len from the event instead
of using preallocated buffer len.

Change-Id: I9e159708be40af172ef78fed39278a1defb40dbd
Signed-off-by: Nikhil Prakash V <nprakash@codeaurora.org>