mhi: Moved writing instance id a step ahead
With pine SoC, writing ERRDBG2 with qrtr instance id
soon after mission ROM image is loaded is not catching
up with the mission ROM execution, hence ERRDBG2 is not
populated with the right qrtr instance id when needed.

To avoid this, moving a step ahead and doing it just
after sbl loading is complete.

Change-Id: I53354c134a6eeb6ffb87539f7c7ca759cbfcc635
Signed-off-by: Gokul Sriram Palanisamy <gokulsri@codeaurora.org>