ipq8074a + SDX55 : Increase initial link retries
This change increase the number of dynamic retries
which are performed during PCIE enumeration in the
boot. This is needed because of  variability in
certain HW boards.

The count essentially sets the max time the boot
code will wait for PCIE enumeration.

Addtionally, this change retains a use delay flag
which triggers a 10 seconds wait on each retry.
The 10 seconds wait method has been used for HW
which don't have wake interrupt mechanism working.
Hence this mechanism has been retained for now.

Change-Id: I3e18e3181710f589f01546544478d56e890b92c7
Signed-off-by: Hardik Desai <hardikd@codeaurora.org>