Panic on receiving invalid packets on IPCR channel.
The host and device maintains its own copy of read and write pointer.
During the ring initialization, the write and read pointers are set
to the base address of the ring.

The host updates the write pointer when it adds an element to the
ring. The doorbell register would contain the position of the
write pointer which the device processes.

The device uses read pointer to determine from where the elements
should be fetched. Once it processes an element from the ring,
it will update the read pointer. After that the device sends
completion event with the updated read pointer. The host uses this
read pointer obtained from the event ring for its parsing.

However here we could see that this read pointer updated by the device
to be corrupt. It is even less than the transfer ring base.

So panic on receiving such invalid packets on IPCR channel.

Change-Id: I1a59062e2d994153d74f2c03a06d2d335036a0da
Signed-off-by: Nikhil Prakash V <nprakash@codeaurora.org>