mhi: Trigger mhi graceful shutdown only when mdm is alive
Currently the mhi unwind sequence tries to clean up
mhi stack and initiates a graceful shutdown over the
mhi bus to notify the mhi endpoint. But graceful shutdown
wont work when the endpoint is already dead. So perform
a graceful shutdown only when the endpoint is alive

While here, rename the property mhi,use-panic-notifier
to mhi,ssr-negotiate to reflect what we are
doing more precisely.

Change-Id: I050662d6670cb311b9443f25620f377ec368b5fe
Signed-off-by: Sricharan R <sricharan@codeaurora.org>