ipq807x: diag: do not close TYPE_CMD channel
There is an issue while doing wifi down/up that commands not being sent
from QXDM to Q6 subsystem.
It is actually reaching till diag_socket_write but because of cleaning
CMD type during wifi down enough information was not available to pass
further. Hence the it gets failed.

With this patch we are not releaseing  the CMD channels.

Change-Id: Ida0dd62f544c9533a2a5ac8825f97473e84cf53b
Signed-off-by: Tushar Nimkar <tnimkar@codeaurora.org>