net: cnss2: Fix memory leak in BDF Download to FW.
In cnss_wlfw_bdf_dnld_send_sync request_firmware allocates memory
to fw_entry but it is released only in error case. In success case,
the memory is not released and new memory is allocated again on
next call of cnss_wlfw_bdf_dnld_send_sync.

Fix this by adding release_firmware in success case if fw_entry is
not NULL.

Change-Id: I54c0735d7f1e2ded69ef756c78966078b6a0ea3f
Signed-off-by: Vignesh Viswanathan <viswanat@codeaurora.org>