mtd: nand: Use single addr cycle for getting nand param
For NAND_CMD_PARAM command, which is used for reading
flash params, 0xEC command id along with address 0x0
will be given to the QPIC controller. In HPG, it is
mentioned that for address 0x0, it is enough to use
single address cycle. So, added support to configure
the number of address cycles, for NAND_CMD_PARAM command,
using dts. Enabling this change only for ipq6018 for now.

Change-Id: I2f8dc61f149bc4343c2109481e803688fb22a000
Signed-off-by: Balaji Prakash J <bjagadee@codeaurora.org>