ipq: mhi: mhitest driver: MHI_STATE_RESET verify
Issue is seen while going for recovery post rddm collection.
mhitest driver in the recovery path, issues the MHI_STATE_RESET and checkes
whether the device's reset bit is cleared or not.

Currently mhitest driver does this little late or post iounmap of BAR
which cause kernel not able to handle paging request.

So this patch solves this issue and does the reset check soon issue
MHI_STATE_RESET or before iounmap BAR.

Change-Id: I925e63d44ef4d3d733a7083b33d2bc7502c851a2
Signed-off-by: Tushar Nimkar <tnimkar@codeaurora.org>