driver: nand: nand_base: Fix multipage read issue
This change will fix multipage read issue. In multipage read
if column address is more means the length requested from
file system will be splitted in multiple page read. In multi page
read we are calculating number of pages and reading one page or
multiple page at a time and subtracting number of pages from total page
read.

If length is less than page size then after first read itself the number
of pages becoming zero and then -1 and so on. and this is messing up
with subsequent read request.

Signed-off-by: Md Sadre Alam <mdalam@codeaurora.org>
Change-Id: I6e6df9f2a539974a9a03a86ca82f050fbd845320