arm64: dts: ipq5018: Update thermal nodes
In ipq5018 platform, 5 sensors are used and sensor ids are mapped
1-1 unlike in ipq807x platform. This patch updates the same.

This patch also adds qfprom nvmem node because tsens calibration data
for ipq5018 will be read from qfprom and removes the int quirk boolean
which is currently unused in ipq5018 platform.

Note that, this quirk might be added later for ipq5018 platform as
and when needed.

Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>
Change-Id: Ice93c0866fa5e95efd70c61839e594057a919151