net: cnss2: Set target_asserted flag for Fatal/Sys Error from MHI
Currently target_asserted flag in plat_priv is set for Fatal Error
and RDDM notification. In case Sys Error is received and RDDM
notification is not received before the timeout, then the
target_asserted flag would not be set leading to issues in
Triage scripts.

Set the target_asserted flag and target_assert_timestamp for
Fatal and Sys Error notifications from MHI

Signed-off-by: Vignesh Viswanathan <viswanat@codeaurora.org>
Change-Id: I9dd3a834e691e8bfcb6bee47b201fe4ba13c993e