arm64: dts: ipq5018: update gpio value for console uart
This change will update gpio config value for console uart
for DB-MP03.1 rdp.

For DB-MP03.1 we have to set GPIO28 & GPIO29 to 0xCC and also
we have to configure GPIO20 & GPIO21 as simple gpio with pull down.
i.e 0xC1.

GPIO20 & GPIO21 is for uart console for AP-MP03.1 & AP-MP03.3 rdps.

Signed-off-by: Md Sadre Alam <mdalam@codeaurora.org>
Change-Id: Ifce4f5f530d7c11a6e59a6c47fe18a3d1b64f43d