driver: nand: nand_base: Fix ubi error in multipage read.
This change will fix ubifs error while reading in multipage read.

In multipage read if if col is not 0 then data buffer is getting
over written due to that upper layer is seeing the data curroption.

so added condition to check for col value is 0 or not.

if (col != 0) {
	req_pages = 1;
	bytes = min(mtd->writesize - col, readlen);
	aligned = (bytes == mtd->writesize);
}

error:
[  123.810139] UBIFS error (ubi0:4 pid 660): ubifs_decompress:
cannot decompress 3167 bytes, compressor lzo, error -22
[  123.814208] UBIFS error (ubi0:4 pid 660): read_block: bad data node
(block 52, inode 310)

Signed-off-by: Md Sadre Alam <mdalam@codeaurora.org>
Change-Id: I9c77ae10fbc73eb9aa3d2a5279fb523480503a4b