clk: ipq5018: Add IGNORE_UNUSED flag for GMAC clks
This change is to add 'CLK_IGNORE_UNUSED' flag for
snoc_gmac0_axi and snoc_gmac1_axi clocks, which are currenlty
enabled during BOOT and crucial for Ethernet transactions.

The flags will be removed later to allow disbaling these
unused clocks during initial kernel boot-up and are expected
to be enabled later by the respective program.

Signed-off-by: Pavithra Palanisamy <pavip@codeaurora.org>
Change-Id: I30f8b1199287067d9ad0d194490c6083ea0cacfa